Changed App.js to fetch news and pass it as props
